+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Fixed
+- The web version now has sound on plain HTTP
+
 ## [0.19.0] - 2026-09-06
 
 > **One more Android tester needed!**